use entry menu for entry list context menu closes #1640
