+++ frontend/viewer/src/project/browse/EntryMenu.svelte
@@ -3,22 +3,41 @@
   import { type IconClass } from '$lib/icon-class';
   import * as DropdownMenu from '$lib/components/ui/dropdown-menu';
   import * as Drawer from '$lib/components/ui/drawer';
+  import * as ContextMenu from '$lib/components/ui/context-menu';
   import { IsMobile } from '$lib/hooks/is-mobile.svelte';
   import { t } from 'svelte-i18n-lingui';
   import { buttonVariants } from '$lib/components/ui/button';
   import Button from '$lib/components/ui/button/button.svelte';
   import {useMultiWindowService} from '$lib/services/multi-window-service';
   import type {IEntry} from '$lib/dotnet-types';
+  import type {Snippet} from 'svelte';
+  import {useWritingSystemService} from '$lib/writing-system-service.svelte';
+  import {useDialogsService} from '$lib/services/dialogs-service';
+  import {useProjectEventBus} from '$lib/services/event-bus';
+  import {useMiniLcmApi} from '$lib/services/service-provider';
 
   const multiWindowService = useMultiWindowService();
+  const dialogsService = useDialogsService();
+  const projectEventBus = useProjectEventBus();
+  const writingSystemService = useWritingSystemService();
+  const miniLcmApi = useMiniLcmApi();
 
-  let { onDelete, entry } = $props<{
-    onDelete?: () => void;
+  let { entry, contextMenu = false, children = undefined } = $props<{
     entry: IEntry;
+    contextMenu?: boolean;
+    children?: Snippet
   }>();
+
+  const headword = $derived((entry && writingSystemService.headword(entry)) || $t`Untitled`);
 
   let open = $state(false);
   const triggerVariant = buttonVariants({ variant: 'ghost', size: 'sm', class: 'float-right' });
+
+  async function onDelete() {
+    if (!await dialogsService.promptDelete($t`Entry`, headword)) return;
+    await miniLcmApi.deleteEntry(entry.id);
+    projectEventBus.notifyEntryDeleted(entry.id);
+  }
 </script>
 
 {#snippet items()}
@@ -30,7 +49,12 @@
 {/snippet}
 
 {#snippet menuItem(icon: IconClass, label: string, onSelect: () => void)}
-  {#if !IsMobile.value}
+  {#if contextMenu}
+    <ContextMenu.Item class="cursor-pointer" onclick={onSelect}>
+      <Icon icon={icon} class="mr-2"/>
+      {label}
+    </ContextMenu.Item>
+  {:else if !IsMobile.value}
     <DropdownMenu.Item class="cursor-pointer" {onSelect}>
       <Icon {icon} class="mr-2" />
       {label}
@@ -46,8 +70,16 @@
     </Button>
   {/if}
 {/snippet}
-
-{#if !IsMobile.value}
+{#if contextMenu}
+  <ContextMenu.Root>
+    <ContextMenu.Trigger>
+      {@render children?.()}
+    </ContextMenu.Trigger>
+    <ContextMenu.Content>
+      {@render items()}
+    </ContextMenu.Content>
+  </ContextMenu.Root>
+{:else if !IsMobile.value}
   <DropdownMenu.Root bind:open>
     <DropdownMenu.Trigger class={triggerVariant}>
       <Icon icon="i-mdi-dots-vertical" class="cursor-pointer" />
